+++ rapid_videocr/main.py
@@ -3,24 +3,13 @@
 # @Contact: liekkaskono@163.com
 import argparse
 from pathlib import Path
-from typing import Any, Dict, List, Optional, Tuple, Union
-
-import cv2
-import numpy as np
-from rapidocr import RapidOCR
-from tqdm import tqdm
+from typing import Any, Dict, List, Optional, Union
 
 from .export import ExportStrategyFactory, OutputFormat
+from .ocr_processor import OCRProcessor
 from .utils.crop_by_project import CropByProject
 from .utils.logger import Logger
-from .utils.utils import (
-    compute_centroid,
-    compute_poly_iou,
-    is_inclusive_each_other,
-    mkdir,
-    padding_img,
-    read_img,
-)
+from .utils.utils import mkdir
 
 
 class RapidVideOCR:
@@ -93,205 +82,6 @@ class RapidVideOCR:
     def print_console(self, txt_result: List):
         for v in txt_result:
             print(v.strip())
-
-
-class OCRProcessor:
-    def __init__(self, ocr_params, batch_size: int):
-        self.logger = Logger(logger_name=__name__).get_log()
-
-        default_params = {"Global.width_height_ratio": -1}
-        if ocr_params is not None:
-            ocr_params.update(default_params)
-        else:
-            ocr_params = default_params
-        self.ocr_engine = RapidOCR(params=ocr_params)
-
-        self.batch_size = batch_size
-
-    def __call__(self, img_list: List[Path], is_batch_rec: bool, is_txt_dir: bool):
-        self.is_txt_dir = is_txt_dir
-        process_func = self.batch_rec if is_batch_rec else self.single_rec
-        srt_result, txt_result = process_func(img_list)
-        return srt_result, txt_result
-
-    def single_rec(self, img_list: List[Path]) -> Tuple[List[str], List[str]]:
-        self.logger.info("[OCR] Running with single recognition.")
-
-        srt_result, txt_result = [], []
-        for i, img_path in enumerate(tqdm(img_list, desc="OCR")):
-            time_str = self.get_srt_timestamp(img_path)
-
-            img = read_img(img_path)
-            img = padding_img(img, (img.shape[0], img.shape[0], 0, 0))
-
-            dt_boxes, rec_res = self.get_ocr_result(img)
-            if dt_boxes is None or rec_res is None:
-                txts = ""
-            else:
-                txts = self.process_same_line(dt_boxes, rec_res)
-
-            srt_result.append(f"{i+1}\n{time_str}\n{txts}\n")
-            txt_result.append(f"{txts}\n")
-        return srt_result, txt_result
-
-    @staticmethod
-    def get_srt_timestamp(file_path: Path) -> str:
-        split_paths = file_path.stem.split("_")
-        start_time = split_paths[:4]
-        start_time[0] = f"{start_time[0]:0>2}"
-        start_str = ":".join(start_time[:3]) + f",{start_time[3]}"
-
-        end_time = split_paths[5:9]
-        end_time[0] = f"{end_time[0]:0>2}"
-        end_str = ":".join(end_time[:3]) + f",{end_time[3]}"
-        return f"{start_str} --> {end_str}"
-
-    def batch_rec(self, img_list: List[Path]) -> Tuple[List, List]:
-        self.logger.info("[OCR] Running with concat recognition.")
-
-        img_nums = len(img_list)
-        srt_result, txt_result = [], []
-        for start_i in tqdm(range(0, img_nums, self.batch_size), desc="Concat Rec"):
-            end_i = min(img_nums, start_i + self.batch_size)
-
-            concat_img, img_coordinates, img_paths = self.get_batch_img_info(
-                img_list, start_i, end_i
-            )
-            dt_boxes, rec_res = self.get_ocr_result(concat_img)
-            if rec_res is None or dt_boxes is None:
-                continue
-
-            srt_part, txt_part = self.get_match_results(
-                start_i, img_coordinates, dt_boxes, rec_res, img_paths
-            )
-            srt_result.extend(srt_part)
-            txt_result.extend(txt_part)
-        return srt_result, txt_result
-
-    def get_batch_img_info(
-        self, img_list: List[Path], start: int, end: int
-    ) -> Tuple[np.ndarray, np.ndarray, List[Path]]:
-        select_imgs = img_list[start:end]
-
-        padding_value = 10
-        array_img_list, img_coordinates, batch_img_paths = [], [], []
-        for i, img_path in enumerate(select_imgs):
-            batch_img_paths.append(img_path)
-
-            img = read_img(img_path)
-
-            if self.is_txt_dir:
-                img = cv2.resize(img, None, fx=0.25, fy=0.25)
-
-            pad_img = padding_img(img, (0, padding_value, 0, 0))
-            array_img_list.append(pad_img)
-
-            h, w = img.shape[:2]
-            x0, y0 = 0, i * (h + padding_value)
-            x1, y1 = w, (i + 1) * (h + padding_value)
-            img_coordinates.append([(x0, y0), (x1, y0), (x1, y1), (x0, y1)])
-
-        concat_img = np.vstack(array_img_list)
-        return concat_img, np.array(img_coordinates), batch_img_paths
-
-    def get_match_results(
-        self,
-        start_i: int,
-        img_coordinates: np.ndarray,
-        dt_boxes: np.ndarray,
-        rec_res: Tuple[str],
-        img_paths: List[Path],
-    ) -> Tuple[List[str], List[str]]:
-        match_dict: Dict[int, List[Union[Path, np.ndarray, str]]] = {
-            k: [] for k in range(len(img_coordinates))
-        }
-        visited_idx = []
-        for i, frame_boxes in enumerate(img_coordinates):
-            for idx, dt_box in enumerate(dt_boxes):
-                if idx in visited_idx:
-                    continue
-
-                box_iou = compute_poly_iou(frame_boxes, dt_box)
-                if is_inclusive_each_other(frame_boxes, dt_box) or box_iou > 0.1:
-                    matched_path = img_paths[i]
-                    match_dict.setdefault(i, []).append(
-                        [matched_path, dt_box, rec_res[idx]]
-                    )
-                    visited_idx.append(idx)
-
-        srt_result_part, txt_result_part = [], []
-        for k, v in match_dict.items():
-            cur_frame_idx = start_i + k
-            if v:
-                img_path, boxes, recs = list(zip(*v))
-                time_str = self.get_srt_timestamp(img_path[0])
-                txts = self.process_same_line(boxes, recs)
-            else:
-                time_str = self.get_srt_timestamp(img_paths[k])
-                txts = ""
-
-            srt_result_part.append(f"{cur_frame_idx+1}\n{time_str}\n{txts}\n")
-            txt_result_part.append(f"{txts}\n")
-        return srt_result_part, txt_result_part
-
-    def get_ocr_result(
-        self, img: np.ndarray
-    ) -> Tuple[Optional[np.ndarray], Optional[Tuple[str]]]:
-        ocr_result = self.ocr_engine(img)
-        if ocr_result.boxes is None:
-            return None, None
-        return ocr_result.boxes, ocr_result.txts
-
-    def process_same_line(self, dt_boxes: np.ndarray, rec_res: List[str]) -> str:
-        """处理同一行文本的识别结果
-
-        Args:
-            dt_boxes: 检测框坐标数组
-            rec_res: 识别结果列表
-
-        Returns:
-            合并后的文本结果字符串
-        """
-        if len(rec_res) == 1:
-            return rec_res[0]
-
-        y_centroids = [compute_centroid(box)[1] for box in dt_boxes]
-        line_groups = self._group_by_lines(y_centroids)
-        return self._merge_line_text(line_groups, rec_res)
-
-    def _group_by_lines(self, y_centroids: List[float]) -> List[List[int]]:
-        """将文本框按行分组"""
-        bool_res = self.is_same_line(y_centroids)
-        groups = []
-        current_group = [0]
-        for i, is_same in enumerate(bool_res, 1):
-            if is_same:
-                current_group.append(i)
-            else:
-                groups.append(current_group)
-                current_group = [i]
-
-        groups.append(current_group)
-        return groups
-
-    def _merge_line_text(self, line_groups: List[List[int]], rec_res: List[str]) -> str:
-        lines = []
-        for group in line_groups:
-            line_text = " ".join(rec_res[i] for i in group)
-            lines.append(line_text)
-        return "\n".join(lines)
-
-    @staticmethod
-    def is_same_line(points: List) -> List[bool]:
-        threshold = 5
-
-        align_points = list(zip(points, points[1:]))
-        bool_res = [False] * len(align_points)
-        for i, point in enumerate(align_points):
-            y0, y1 = point
-            if abs(y0 - y1) <= threshold:
-                bool_res[i] = True
-        return bool_res
 
 
 class RapidVideOCRExeception(Exception):
